#403 Split decorators into multiple files
