+++ twake/backend/node/src/core/platform/framework/decorators/consumes.ts
@@ -1,0 +1,8 @@
+import { CONSUMES_METADATA } from "../api";
+
+export function Consumes(services: string[] = []): ClassDecorator {
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  return function(target: Function): void {
+    Reflect.defineMetadata(CONSUMES_METADATA, services, target.prototype);
+  };
+}
