+++ twake/backend/node/src/core/platform/framework/decorators/index.ts
@@ -1,0 +1,3 @@
+export * from "./consumes";
+export * from "./prefix";
+export * from "./service-name";
