+++ twake/backend/node/src/core/platform/framework/decorators/prefix.ts
@@ -1,0 +1,8 @@
+import { PREFIX_METADATA } from "../api";
+
+export function Prefix(prefix: string = "/"): ClassDecorator {
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  return function(target: Function): void {
+    Reflect.defineMetadata(PREFIX_METADATA, prefix, target.prototype);
+  };
+}
